+++ packages/plugin-ext/src/main/browser/debug/plugin-debug-session-factory.ts
@@ -34,6 +34,7 @@ export class PluginDebugSession extends DebugSession {
     constructor(
         readonly id: string,
         readonly options: DebugSessionOptions,
+        readonly parentSession: DebugSession | undefined,
         protected readonly connection: DebugSessionConnection,
         protected readonly terminalServer: TerminalService,
         protected readonly editorManager: EditorManager,
@@ -42,7 +43,7 @@ export class PluginDebugSession extends DebugSession {
         protected readonly messages: MessageClient,
         protected readonly fileService: FileService,
         protected readonly terminalOptionsExt: TerminalOptionsExt | undefined) {
-        super(id, options, connection, terminalServer, editorManager, breakpoints, labelProvider, messages, fileService);
+        super(id, options, parentSession, connection, terminalServer, editorManager, breakpoints, labelProvider, messages, fileService);
     }
 
     protected async doCreateTerminal(terminalWidgetOptions: TerminalWidgetOptions): Promise<TerminalWidget> {
@@ -71,7 +72,7 @@ export class PluginDebugSessionFactory extends DefaultDebugSessionFactory {
         super();
     }
 
-    get(sessionId: string, options: DebugSessionOptions): DebugSession {
+    get(sessionId: string, options: DebugSessionOptions, parentSession?: DebugSession): DebugSession {
         const connection = new DebugSessionConnection(
             sessionId,
             this.connectionFactory,
@@ -80,6 +81,7 @@ export class PluginDebugSessionFactory extends DefaultDebugSessionFactory {
         return new PluginDebugSession(
             sessionId,
             options,
+            parentSession,
             connection,
             this.terminalService,
             this.editorManager,